requirements.txt
flask
numpy
python-dotenv
openai